requirements.txt
accelerate==1.7.0          		aiohappyeyeballs==2.6.1   		aiohttp==3.12.9
aiosignal==1.3.2           		async-timeout==5.0.1      		attrs==25.3.0
bitsandbytes==0.46.0       		certifi==2025.4.26        		charset-normalizer==3.4.2
datasets==3.6.0            		-e git+https://github.com/huggingface/diffusers.git@73a9d5856f2d7ae3637c484d83cd697284ad3962#egg=diffusers   
dill==0.3.8
filelock==3.18.0           		frozenlist==1.6.2         		fsspec==2025.3.0
hf-xet==1.1.3              		huggingface-hub==0.32.4   		idna==3.10
importlib_metadata==8.7.0  		Jinja2==3.1.6             		MarkupSafe==3.0.2
mpmath==1.3.0              		multidict==6.4.4          		multiprocess==0.70.16
networkx==3.4.2            		numpy==2.2.6              		nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80   	nvidia-cuda-nvrtc-cu12==12.6.77  	nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17       	nvidia-cufft-cu12==11.3.0.4      	nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77     	nvidia-cusolver-cu12==11.7.1.2   	nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3     	nvidia-nccl-cu12==2.26.2         	nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77         	packaging==25.0                  	pandas==2.3.0
peft==0.15.2              		pillow==11.2.1             		propcache==0.3.1
psutil==7.0.0             		pyarrow==20.0.0            		python-dateutil==2.9.0.post0
pytz==2025.2              		PyYAML==6.0.2              		regex==2024.11.6
requests==2.32.3          		safetensors==0.5.3         		six==1.17.0
sympy==1.14.0             		tokenizers==0.21.1         		torch==2.7.1
torchvision==0.22.1       		tqdm==4.67.1               		transformers==4.52.4
triton==3.3.1             		typing_extensions==4.14.0  		tzdata==2025.2
urllib3==2.4.0            		xxhash==3.5.0              		yarl==1.20.0
zipp==3.22.0
